Move feature sub folders out of component dir.
