+++ coral/src/app/features/requests/RequestResourcesTabs.test.tsx
@@ -1,6 +1,6 @@
 import { cleanup, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import RequestsResourceTabs from "src/app/features/requests/components/RequestsResourceTabs";
+import RequestsResourceTabs from "src/app/features/requests/RequestsResourceTabs";
 import { RequestsTabEnum } from "src/app/router_utils";
 import { customRender } from "src/services/test-utils/render-with-wrappers";
