+++ coral/src/app/features/requests/acls/AclRequests.test.tsx
@@ -5,7 +5,7 @@ import {
   waitForElementToBeRemoved,
 } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { AclRequests } from "src/app/features/requests/components/acls/AclRequests";
+import { AclRequests } from "src/app/features/requests/acls/AclRequests";
 import { getAclRequests } from "src/domain/acl/acl-api";
 import transformAclRequestApiResponse from "src/domain/acl/acl-transformer";
 import { getEnvironments } from "src/domain/environment";
