+++ coral/src/app/features/requests/topics/TopicRequests.test.tsx
@@ -7,7 +7,7 @@ import {
 import { getTopicRequests } from "src/domain/topic/topic-api";
 import { transformGetTopicRequestsResponse } from "src/domain/topic/topic-transformer";
 import { mockIntersectionObserver } from "src/services/test-utils/mock-intersection-observer";
-import { TopicRequests } from "src/app/features/requests/components/topics/TopicRequests";
+import { TopicRequests } from "src/app/features/requests/topics/TopicRequests";
 import { customRender } from "src/services/test-utils/render-with-wrappers";
 import userEvent from "@testing-library/user-event";
