+++ coral/src/app/pages/requests/topics/index.tsx
@@ -1,4 +1,4 @@
-import { TopicRequests } from "src/app/features/requests/components/topics/TopicRequests";
+import { TopicRequests } from "src/app/features/requests/topics/TopicRequests";
 
 const TopicRequestsPage = () => {
   return <TopicRequests />;
